Explicitly set region on endpoint-create
Currently, endpoint is getting created as region 'regionOne' (hard-coded
default in keystoneclient).  This may differ from the region we have in
our openrc file and therefore doing something like and endpoint-get
may fail since the region we pass in via environment may not match the
region on the endpoint.

This does not appear to be an issue on older versions of keystoneclient.